cleanup: remove old command files after refactoring
Remove obsolete command files that have been consolidated into the new
analyze/maintain pattern:

- Deleted old analyze.py, clean.py, health.py, merge.py
- Renamed analyze_new.py to analyze.py
- Updated imports in main.py
- Kept gc.py as it contains GarbageCollector class used by maintain.py

This completes the command consolidation refactoring.
